requirements.txt
Django>=4.0.1,<4.1.0
djangorestframework>=3.13.1,<3.14.0
redis>=4.1.1,<4.2.0
django-redis>=5.2.0,<5.3.0
